requirements.txt
langchain
langchain-community
langchain-huggingface
openai 
python-dotenv 
ipykernel
youtube-transcript-api
pytube
tiktoken
llama-index
sentence_transformers
pdf2image
pdfminer
pikepdf
pypdf
PyPDF2
pdfminer.six
unstructured
faiss-cpu
opencv-python
unstructured_pytesseract
unstructured_inference
streamlit
transformers
torch
huggingface-hub